+++ lib/src/gatekeeper_mock.dart
@@ -5,8 +5,16 @@ import 'gatekeeper_base.dart';
 /// system resources like iptables.
 class GatekeeperMock extends GatekeeperDriver {
   final Set<int> blockedPorts;
+  final Set<(String, int)> acceptedAddressesOnPort;
 
-  GatekeeperMock(this.blockedPorts);
+  final bool verbose;
+
+  GatekeeperMock(
+      {Set<int>? blockedPorts,
+      Set<(String, int)>? acceptedAddressesOnPort,
+      this.verbose = false})
+      : blockedPorts = blockedPorts ?? {},
+        acceptedAddressesOnPort = acceptedAddressesOnPort ?? {};
 
   @override
   Future<String> resolveBinaryPath(String binaryCommand) async {
@@ -22,7 +30,13 @@ class GatekeeperMock extends GatekeeperDriver {
   @override
   Future<Set<int>> listBlockedTCPPorts(
       {bool sudo = false, Set<int>? allowedPorts}) async {
-    return blockedPorts.toSet();
+    var set = blockedPorts.toSet();
+
+    if (allowedPorts != null) {
+      set.removeWhere((p) => !allowedPorts.contains(p));
+    }
+
+    return set;
   }
 
   @override
@@ -58,6 +72,51 @@ class GatekeeperMock extends GatekeeperDriver {
     }
 
     blockedPorts.remove(port);
+    return true;
+  }
+
+  @override
+  Future<Set<(String, int)>> listAcceptedAddressesOnTCPPorts(
+      {bool sudo = false, Set<int>? allowedPorts}) async {
+    var set = acceptedAddressesOnPort.toSet();
+
+    if (allowedPorts != null) {
+      set.removeWhere((e) => !allowedPorts.contains(e.$2));
+    }
+
+    return set;
+  }
+
+  @override
+  Future<bool> acceptAddressOnTCPPort(String address, int port,
+      {bool sudo = false,
+      required Set<int>? allowedPorts,
+      required bool allowAllPorts}) async {
+    if (!allowAllPorts &&
+        (allowedPorts == null || !allowedPorts.contains(port))) {
+      return false;
+    }
+
+    acceptedAddressesOnPort.add((address, port));
+    return true;
+  }
+
+  @override
+  Future<bool> unacceptAddressOnTCPPort(String address, int? port,
+      {bool sudo = false,
+      required Set<int>? allowedPorts,
+      required bool allowAllPorts}) async {
+    if (port != null &&
+        !allowAllPorts &&
+        (allowedPorts == null || !allowedPorts.contains(port))) {
+      return false;
+    }
+
+    if (port != null) {
+      acceptedAddressesOnPort.remove((address, port));
+    } else {
+      acceptedAddressesOnPort.removeWhere((e) => e.$1 == address);
+    }
     return true;
   }