+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 1.0.2
+
+- `Gatekeeper`, `GatekeeperDriver`, `GatekeeperClient`, `GatekeeperMock`, `GatekeeperIpTables`:
+  - Added `listAcceptedAddressesOnTCPPorts`, `acceptAddressOnTCPPort`, `unacceptAddressOnTCPPort`, `isAcceptedAddressOnPort`.
+
+- `GatekeeperMock`, `GatekeeperIpTables`:
+  - Added option `verbose`.
+
+- `GatekeeperClient`:
+  - `processCommand`:
+    - Added `list accepted`.
+    - Added `list all`.
+    - Added `accept` and `unaccept`
+
+- `gatekeeper_server`:
+  - Improve console logging.
+  - Process:
+    - `list accepted`.
+    - `list all`.
+    - `accept` and `unaccept`
+
 ## 1.0.1
 
 - `GatekeeperServer`: